pyproject.toml
[project]
name = "robot-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "ctkmessagebox>=2.7",
    "customtkinter>=5.2.2",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pillow>=11.0.0",
    "portalocker>=3.0.0",
    "pyinstaller>=6.11.1",
    "requests>=2.32.3",
]

[tool.pyright]
venvPath = "."
venv = ".venv"

